Level classification
This problem of classifying a product into a
particular category is very important to serve various use cases – like, helping search, performing
product matching, providing category specific insights, and so on.
The problem of stamping every product in our catalog into a category is a Multi Label
Hierarchical Classification problem. Here challenge is to build a micro version of this classifier where
you will predict 4 classes.